AP-8036: Reread Keras model if network spec is outdated
This improves backward compatibility.

A network spec is outdated if it contains a tensor spec without an id.
The tensor id is essential for communication between Java and Python, so
we can't just leave it blank. Also, we can't generate the id on Java
side as it's assigned while traversing the Keras network graph.

Also improve some documentation to clarify the side effects that came
along with introducing the tensor id.

Also some misc. improvements.