AP-9168: Fix issue with names and ids in keras learner
If there is no target config for name look for target config with id.
This is a quickfix and we might decide to change the mapping altogether
in the future but for now this suffices.
Changing the mapping might have serious side effects and requires
careful consideration.

AP-9168 (DL Keras: Create nodes for core layers)