AP-9599: Modify SpecInferrer and Materializer to handle collect layers
Collect layers represent no actual Keras layers but are necessary to
facilitate networks with multiple outputs.
The current framework is not designed for this case and we hence have to
treat collect layers specially in some key functionalities of the
framework.

AP-9599 (DL Keras: Can't create networks with multiple outputs with
layer nodes)