AP-10060: Store created configs during configure
Up until now, we only store the converters which is fine for inputs
since that's the only information we need for the execution. However, in
case of targets, we also need the loss function which is part of the
config. Solution: Also store the config.

AP-10060 (DL Keras: Unconfigured keras network learner throws
NullPointerException)