AP-11351: Remove check if outputs have changed
If the output specs are different the configuration can still be valid.
We only have to check if the configured outputs are available and the
converter is compatible. The check if the configured outputs are
available was already there in `configureOutputs` and the check if the
converter is compatible was added.

Note that now the last configured network spec and last incoming network
spec wasn't used anymore and therefore removed.