AP-14412: Use specific DL Python environment
Add a DLPythonAbstractContext which can be extended to have a context
for different PythonCommand preferences. The DLPythonDefaultContext uses
the PythonCommand for the library which is selected as default. The
DLKerasPythonContext uses the PythonCommand selected for Keras. This
context is now used in all Keras specific nodes.

AP-14412 (TensorFlow 2 environment selection/creation on preference page)