Fix tests: Prevent unregistering the training message handler
Randomly workflow tests with the Keras Network Learner node failed
because training data was requested after the training was done. The
requests could not be handled because the handler was unregistered after
the training. To prevent this bug the message handler is not
unregistered anymore.
Note that the Python kernel is not reused and is closed after the
network is written to disc.