AP-12210: Fix stdout/stderr update in DL Python nodes
In most DL Python nodes the standard output and standard error was
overwritten during the node execution. Therefore it was not the expected
output from the python script. This commit adds a method to the
DLPythonNodeModel which allows to easily update the output without
overwriting it.

AP-12210 (DL Python: stdout and stderr view of nodes empty)