AP-17330: Add installation test to network materialization
The check was missing at that location.

Also streamline the exception delivery in the port object. In
particular, remove the message about a possible Keras/TensorFlow
version mismatch. This message proved to be rather confusing to users
in the past and was therefore removed at another location already.

AP-17330 (Deep learning: Installation tests do not use variable-
controlled Python environment)